+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- TAXJAR-34 Ensure tax adjustment does not reduce tax below zero
+
 ## [0.1.8] - 2021-10-21
 
 ### Added